% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.r
\name{backward}
\alias{backward}
\alias{forward}
\title{Backward Elimination/Forward Selection of Model Terms in logistf Models}
\usage{
backward(x, ...)

forward(
  object,
  scope,
  steps = 1000,
  slentry = 0.05,
  trace = TRUE,
  printwork = FALSE,
  pl = TRUE,
  ...
)
}
\arguments{
\item{...}{Further arguments to be passed to methods.}

\item{object}{A fitted logistf model object. To start with an empty model, create a model fit 
with a formula=<y>~1, pl=FALSE. (Replace <y> by your response variable.)}

\item{scope}{The scope of variables to add/drop from the model. If left blank, backward will use 
the terms of the object fit, and forward will use all variables in object$data not identified as 
the response variable. Alternatively, an arbitrary vector of variable names can be given, to allow 
that only some of the variables will be competitively selected or dropped.}

\item{steps}{The number of forward selection/backward elimination steps.}

\item{slentry}{For \code{forward}, the significance level to enter the model.}

\item{trace}{If \code{TRUE}, protocols selection steps.}

\item{printwork}{If \code{TRUE}, prints each working model that is visited by the selection procedure.}

\item{pl}{For forward, computes profile likelihood confidence intervals for the final model if \code{TRUE}.}

\item{slstay}{For \code{backward}, the significance level to stay in the model.}

\item{full.penalty}{If \code{TRUE} penalty is not taken from current model but from start model.}
}
\value{
An updated \code{logistf, flic} or \code{flac} fit with the finally selected model.
}
\description{
These functions provide simple backward elimination/forward selection procedures for logistf models.
}
\details{
The variable selection is simply performed by repeatedly calling add1 or drop1 methods for logistf, 
and is based on penalized likelihood ratio test. It can also properly handle variables that were 
defined as factors in the original data set.
}
\section{Functions}{
\itemize{
\item \code{forward}: Forward Selection
}}

\examples{
data(sex2) 
fit<-logistf(data=sex2, case~1, pl=FALSE) 
fitf<-forward(fit) 

fit2<-logistf(data=sex2, case~age+oc+vic+vicl+vis+dia) 
fitb<-backward(fit2)

}
