% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans-support.R
\name{emm_basis.logistf}
\alias{emm_basis.logistf}
\title{emm basis method for logistf objects}
\usage{
\method{emm_basis}{logistf}(object, trms, xlev, grid, ...)
}
\arguments{
\item{object}{a \code{logistf} object.}

\item{trms}{a model terms object.}

\item{xlev}{a named list of character vectors giving the full set of levels to be assumed for each factor.}

\item{grid}{a data.frame, list or environment (or object coercible by as.data.frame to a data.frame), containing the variables in formula. Neither a matrix nor an array will be accepted.}

\item{...}{additional arguments.}
}
\description{
This function provides a \code{emm_basis} method for \code{logistf} objects. This is required for \code{emmeans} support.
}
