% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas.R
\name{nas}
\alias{nas}
\title{Generate a Basis Matrix for Natural Cubic Splines}
\usage{
nas(x, df = NULL, knots = NULL, intercept = FALSE, Boundary.knots)
}
\arguments{
\item{x}{the predictor variable}

\item{df}{degrees of freedom. If supplied, df - 1 - intercept knots at suitably chosen quantiles of x are set.}

\item{knots}{knots of the spline.}

\item{intercept}{if \code{TRUE}, an intercept is included in the basis}

\item{Boundary.knots}{boundary points at which to impose the natural boundary conditions and anchor the B-spline basis}
}
\value{
A matrix of dimension \code{length(x) * df} where either \code{df} was supplied or if knots were supplied, 
\code{df = length(knots) + 1 + intercept}.
}
\description{
Generate the B-spline basis matrix for a natural cubic spline. This function calls [splines::ns()] but sets the boundary knots, 
if not given, according to the number of knots and to the guidelines of [Hmisc::rcspline.eval()]. 
For further information please see the individual documentations.
}
\examples{

data(sex2)
sex2$agec <- runif(nrow(sex2), 18, 25)
nas(sex2$agec, df = 3)
summary(fit <- logistf(case ~ nas(sex2$agec, df = 3), data = sex2))
}
\seealso{
[splines::ns()],[Hmisc::rcspline.eval()]
}
