% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.flac.R
\name{predict.flac}
\alias{predict.flac}
\title{Predict Method for flac Fits}
\usage{
\method{predict}{flac}(flicobject, newdata, type = c("link", "response"))
}
\arguments{
\item{newdata}{Optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{type}{The type of prediction required. The default is on the scale of the linear predictors. The alternative \code{response} gives the predicted probabilities.}

\item{lfobject}{A fitted object of class \code{flac}.}
}
\value{
A vector or matrix of predictions.
}
\description{
Obtains predictions from a fitted \code{flac} object.
}
\details{
If \code{newdata} is omitted the predictions are based on the data used for the fit.
}
