% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.flic.R
\name{predict.flic}
\alias{predict.flic}
\title{Predict Method for flic Fits}
\usage{
\method{predict}{flic}(flicobject, newdata, type = c("link", "response"))
}
\arguments{
\item{newdata}{Optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{type}{The type of prediction required. The default is on the scale of the linear predictors. The alternative \code{response} gives the predicted probabilities.}

\item{lfobject}{A fitted object of class \code{flic}.}
}
\value{
A vector or matrix of predictions
}
\description{
Obtains predictions from a fitted \code{flic} object.
}
\details{
If \code{newdata} is omitted the predictions are based on the data used for the fit.
}
