% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.logistf.R
\name{predict.logistf}
\alias{predict.logistf}
\title{Predict Method for logistf Fits}
\usage{
\method{predict}{logistf}(lfobject, newdata, type = c("link", "response"), flic = FALSE)
}
\arguments{
\item{lfobject}{A fitted object of class \code{logistf}.}

\item{newdata}{Optionally, a data frame in which to look for variables with which to predict. 
If omitted, the fitted linear predictors are used.}

\item{type}{The type of prediction required. The default is on the scale of the linear predictors. 
The alternative \code{response} gives the predicted probabilities.}

\item{flic}{If \code{TRUE}(default = \code{FALSE}), predictions are computed with intercept correction.}
}
\value{
A vector or matrix of predictions.
}
\description{
Obtains predictions from a fitted \code{logistf} object.
}
\details{
If \code{newdata} is omitted the predictions are based on the data used for the fit.
}
