\name{print.logistf}
\Rdversion{1.1}
\alias{print.logistf}
\title{
print method for logistf objects
}
\description{
provides the print method for logistf objects}
\usage{
\method{print}{logistf}(x, ...)
}
\arguments{
  \item{x}{a logistf object}
  \item{\dots}{additional parameters}
}
\value{
A formatted table with results from logistf analysis.
}
\author{Georg Heinze}
\keyword{ models }
\keyword{ regression }% __ONLY ONE__ keyword per line
