\name{print.logistftest}
\Rdversion{1.1}
\alias{print.logistftest}
\title{\code{print} method for \code{logistftest} objects}
\description{
Provides the print method for \code{logistftest} objects.}
\usage{
\method{print}{logistftest}(x, ...)
}
\arguments{
  \item{x}{a logistftest object}
  \item{\dots}{additional parameters}
}
\value{
Side effect: prints results from a penalized likelihood ratio test performed by \code{logistftest}.
}
\author{Georg Heinze}
\keyword{ models }
\keyword{ regression}% __ONLY ONE__ keyword per line
