% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans-support.R
\name{recover_data.logistf}
\alias{recover_data.logistf}
\title{Recover data method for logistf objects}
\usage{
\method{recover_data}{logistf}(object, frame = object$model, ...)
}
\arguments{
\item{object}{a \code{logistf} object.}

\item{frame}{the model frame.}

\item{...}{additional arguments.}
}
\description{
This function provides a \code{recover_data} method for \code{logistf} objects. This is required for \code{emmeans} support.
}
