\name{sexagg}
\alias{sexagg}
\docType{data}
\title{Condom use and first time urinary tract infection
}
\description{
The case-control study of Foxman et al (1997) examines urinary tract
infection in related to age and contraceptive use.  The data set
consists of 130 college women with urinary tract infections and 109
uninfected controls. The data set include the binary covariates age
(age), oral contraceptive use (oc), condom use (vic), lubricated condom
use (vicl), spermicide use (vis) and diaphragm used (dia).
}

\usage{data(sexagg)}

\format{A data frame table.}

\references{
Foxman B, Marsh J, Gillespie B, Rubin N, Koopman JS, Spear S (1997).
  Condom Use and First-Time Urinary Tract Infection. \emph{Epidemiology}
  8: 637--641.
}

\examples{
data(sexagg)
}

\keyword{datasets}
