% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flac.R
\name{flac}
\alias{flac}
\alias{flac.formula}
\alias{flac.logistf}
\title{FLAC - Firth's logistic regression with added covariate}
\usage{
flac(x, ...)

\method{flac}{formula}(formula = attr(data, "formula"), data = sys.parent())

\method{flac}{logistf}(lfobject)
}
\arguments{
\item{x}{Either formula and data or \code{\link{logistf}} object}

\item{...}{Further arguments passed to the method.}
}
\value{
A \code{flac} object with components:
  \item{coefficients}{The coefficients of the parameter in the fitted model.}
  \item{predicted.probabilities}{A vector with the predicted probability of each observation}
  \item{linear.predictions}{A vector with the linear predictor of each observation.}
  \item{probabilities}{The p-values of the specific parameters}
  \item{ci.lower}{The lower confidence limits of the parameter.}
  \item{ci.upper}{The upper confidence limits of the parameter.}
  \item{call}{The call object.}
  \item{alpha}{The significance level: 0.95}
  \item{method}{The fitting method: 'Penalized ML'}
  \item{method.ci}{The method calculating the confidence intervals of the parameter (without intercept.)}
  \item{var}{The variance-covariance-matrix of the parameters.}
}
\description{
\code{flac} implements Firth's bias-Reduced penalized-likelihood logistic regression with added covariate.
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: With formula and data

\item \code{logistf}: With logistf object
}}

\examples{
#With formula and data:
data(sex2)
flac(case ~ age + oc + vic + vicl + vis + dia, sex2)

#With a logistf object:
lf <- logistf(formula = case ~ age + oc + vic + vicl + vis + dia, data = sex2)
flac(lf)

}
\seealso{
[logistf()] for Firth's bias-Reduced penalized-likelihood logistic regression.
}
