% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flic.R
\name{flic}
\alias{flic}
\alias{flic.formula}
\alias{flic.logistf}
\title{FLIC - Firth's logistic regression with intercept correction}
\usage{
flic(x, ...)

\method{flic}{formula}(formula = attr(data, "formula"), data = sys.parent())

\method{flic}{logistf}(lfobject)
}
\arguments{
\item{x}{Either formula and data or \code{\link{logistf}} object}

\item{...}{Further arguments passed to the method.}
}
\value{
A \code{flic} object with components:
  \item{coefficients}{The coefficients of the parameter in the fitted model.}
  \item{predicted.probabilities}{A vector with the predicted probability of each observation}
  \item{linear.predictions}{A vector with the linear predictor of each observation.}
  \item{probabilities}{The p-values of the specific parameters}
  \item{ci.lower}{The lower confidence limits of the parameter.}
  \item{ci.upper}{The upper confidence limits of the parameter.}
  \item{call}{The call object.}
  \item{alpha}{The significance level: 0.95}
  \item{method}{The fitting method: 'Penalized ML'}
  \item{method.ci}{The method calculating the confidence intervals of the parameter (without intercept.)}
  \item{var}{The variance-covariance-matrix of the parameters.}
}
\description{
\code{flic} implements Firth's bias-Reduced penalized-likelihood logistic regression with intercept correction.
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: With formula and data

\item \code{logistf}: With logistf object
}}

\examples{
#With formula and data:
data(sex2)
flic(case ~ age + oc + vic + vicl + vis + dia, sex2)

#With a logistf object:
lf <- logistf(formula = case ~ age + oc + vic + vicl + vis + dia, data = sex2)
flic(lf)

}
\seealso{
\code{\link{logistf}} for Firth's bias-Reduced penalized-likelihood logistic regression.
}
