% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sex2}
\alias{sex2}
\title{Urinary Tract Infection in American College Students}
\format{
sex2: a data.frame containing 239 observations
}
\source{
\href{www.cytel.com}
The full data was available for download at their webpage, and is distributed along with their LogXact software. Here a reduced version of the original data is reproduced.
}
\usage{
sex2
}
\description{
This data set deals with urinary tract infection in sexually active college women, along with covariate information on age an contraceptive use. 
The variables are all binary and coded in 1 (condition is present) and 0 (condition is absent): 
case (urinary tract infection, the study outcome variable), age (>= 24 years), dia (use of diaphragm), oc (use of oral contraceptive), vic (use of condom), vicl (use of lubricated condom), and vis (use of spermicide).
}
\references{
Cytel Inc., (2010) LogXact 9 user manual, Cambridge, MA:Cytel Inc
}
\keyword{datasets}
