% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sexagg}
\alias{sexagg}
\title{Urinary Tract Infection in American College Students}
\format{
sexagg: an aggregated data.frame containing 31 observations with case weights (COUNT).
\describe{
\item{case}{urinary tract infection, the study outcome variable}
\item{age}{>= 24 years}
\item{dia}{use of diaphragm}
\item{oc}{use of oral contraceptive}
\item{vic}{use of condom}
\item{vicl}{use of lubricated condom}
\item{vis}{use of spermicide}
}
}
\source{
<http://www.cytel.com/>
}
\usage{
sexagg
}
\description{
This data set deals with urinary tract infection in sexually active college women, 
along with covariate information on age an contraceptive use. 
The variables are all binary and coded in 1 (condition is present) and 0 (condition is absent): 
case (urinary tract infection, the study outcome variable), age (>= 24 years), 
dia (use of diaphragm), oc (use of oral contraceptive), vic (use of condom), 
vicl (use of lubricated condom), and vis (use of spermicide).
}
\references{
Cytel Inc., (2010) LogXact 9 user manual, Cambridge, MA:Cytel Inc
}
\keyword{datasets}
